trying to appease the linter

+++ apis/python/examples/soco-slice-query.py
@@ -30,20 +30,21 @@ def soco_query_and_store(
         print("Empty slice")
         return
 
-    result_soma_slice = tiledbsc.SOMASlices.concat(result_soma_slices)
+    result_soma_slice = tiledbsc.SOMASlice.concat(result_soma_slices)
 
-    if output_h5ad_path is not None:
-        a = result_soma_slice.to_anndata()
-        a.write_h5ad(output_h5ad_path)
-        print("Wrote", output_h5ad_path, a.X.shape)
+    if result_soma_slice is not None:
+        if output_h5ad_path is not None:
+            a = result_soma_slice.to_anndata()
+            a.write_h5ad(output_h5ad_path)
+            print("Wrote", output_h5ad_path, a.X.shape)
 
-    if output_soma_path is not None:
-        if os.path.exists(output_soma_path):
-            shutil.rmtree(output_soma_path)
-        soma = tiledbsc.SOMA.from_soma_slice(result_soma_slice, output_soma_path)
-        data = soma.X.data
-        assert data is not None
-        print("Wrote", output_soma_path, data.shape())
+        if output_soma_path is not None:
+            if os.path.exists(output_soma_path):
+                shutil.rmtree(output_soma_path)
+            soma = tiledbsc.SOMA.from_soma_slice(result_soma_slice, output_soma_path)
+            data = soma.X.data
+            assert data is not None
+            print("Wrote", output_soma_path, data.shape())
 
 
 # ----------------------------------------------------------------
